add: 完善内部结构

+++ common/src/main/java/cn/acitrus/uni/common/entities/base/AbstractUniEntity.java
@@ -49,5 +49,5 @@ public abstract class AbstractUniEntity extends AbstractUniSimpleEntity {
 
     @Version
     @Column
-    public Long version;
+    public Integer version;
 }
